Restore fractional degree tile support from BikeHopper; add changes summary
- Restored double type for tile coordinates in HeightTile and AbstractTiffElevationProvider
- These changes support fractional degree elevation tiles (e.g., 0.25° x 0.25°)
- Updated CGIARProvider and GMTEDProvider to match new signatures
- Added BIKEHOPPER_CHANGES_SUMMARY.md documenting all BikeHopper integration changes
- Note: USGS provider was removed as the data source was discontinued, but infrastructure changes retained for future use

+++ core/src/main/java/com/graphhopper/reader/dem/CGIARProvider.java
@@ -151,12 +151,12 @@ public class CGIARProvider extends AbstractTiffElevationProvider {
     }
 
     @Override
-    int getMinLatForTile(double lat) {
+    double getMinLatForTile(double lat) {
         return down(lat);
     }
 
     @Override
-    int getMinLonForTile(double lon) {
+    double getMinLonForTile(double lon) {
         return down(lon);
     }
